+++ selcukes-databind/src/main/java/io/github/selcukes/databind/converters/LocalDateTimeConverter.java
@@ -14,24 +14,24 @@
  *  limitations under the License.
  */
 
-package io.github.selcukes.excel.converters;
+package io.github.selcukes.databind.converters;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
 
-import static java.time.LocalDate.parse;
+import static java.time.LocalDateTime.parse;
 import static java.time.format.DateTimeFormatter.ofPattern;
 import static java.util.Optional.ofNullable;
 
-public class LocalDateConverter extends DefaultConverter<LocalDate> {
-    private static final String DEFAULT_FORMAT = "yyyy-MM-dd";
+public class LocalDateTimeConverter extends DefaultConverter<LocalDateTime> {
+    public static final String DEFAULT_FORMAT = "yyyy-MM-dd HH:mm:ss";
 
     @Override
-    public LocalDate convert(final String value) {
+    public LocalDateTime convert(final String value) {
         return convert(value, DEFAULT_FORMAT);
     }
 
     @Override
-    public LocalDate convert(final String value, final String format) {
+    public LocalDateTime convert(final String value, final String format) {
         return parse(value, ofPattern(ofNullable(format).filter(f -> !f.isEmpty()).orElse(DEFAULT_FORMAT)));
     }
 }
