+++ selcukes-databind/src/main/java/io/github/selcukes/databind/utils/Streams.java
@@ -1,0 +1,33 @@
+/*
+ *  Copyright (c) Ramesh Babu Prudhvi.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package io.github.selcukes.databind.utils;
+
+import lombok.experimental.UtilityClass;
+
+import java.util.Iterator;
+import java.util.Spliterator;
+import java.util.Spliterators;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
+
+@UtilityClass
+public class Streams {
+    public <T> Stream<T> of(Iterator<? extends T> iterator) {
+        return StreamSupport
+            .stream(Spliterators.spliteratorUnknownSize(iterator, Spliterator.ORDERED), true);
+    }
+}
