+++ selcukes-excel-runner/src/test/java/io/github/selcukes/excel/ExcelMapperTest.java
@@ -16,13 +16,13 @@
 
 package io.github.selcukes.excel;
 
+import io.github.selcukes.databind.DataMapper;
+import io.github.selcukes.databind.ExcelMapper;
+import io.github.selcukes.databind.annotation.Column;
 import io.github.selcukes.databind.annotation.DataFile;
 import io.github.selcukes.databind.exception.DataMapperException;
-import io.github.selcukes.excel.annotation.Column;
 import lombok.Data;
 import org.testng.annotations.Test;
-
-import java.util.stream.Stream;
 
 public class ExcelMapperTest {
 
@@ -35,19 +35,19 @@ public class ExcelMapperTest {
         private String feature;
         @Column(name = "Test")
         private String test;
-       // @Column(name = "Run")
+        // @Column(name = "Run")
         private String run;
     }
 
     @Test
     public void excelMapperTest() {
-        Stream<Pojo> pojoStream = ExcelMapper.parse(Pojo.class);
-        pojoStream.forEach(System.out::println);
+        Pojo pojoStream = DataMapper.parse(Pojo.class);
+        // pojoStream.forEach(System.out::println);
     }
 
     @Test(expectedExceptions = DataMapperException.class)
     public void excelMapperNegativeTest() {
-       ExcelMapper.parse(Selcukes.class);
+        ExcelMapper.parse(Selcukes.class);
 
     }
